fix(sandbox): preserve direct grading entry point
