+++ scripts/thebitlab_runtime_sandbox.py
@@ -8,6 +8,7 @@ from pathlib import Path, PurePosixPath
 from typing import Any, Protocol
 
 from scripts import thebitlab_runtime_plugins
+from scripts.thebitlab_sandbox_boundary import docker_boundary_command
 
 RUNTIME_SANDBOX_RESULT_SCHEMA_VERSION = "runtime_sandbox_result.v1"
 DEFAULT_RUNTIME_SANDBOX_TIMEOUT_GRACE_SECONDS = 10
@@ -21,55 +22,6 @@ class RuntimeSandboxExecutionService(Protocol):
         plan: thebitlab_runtime_plugins.RuntimeSandboxPlan,
         request: thebitlab_runtime_plugins.RuntimeRequest,
     ) -> dict[str, Any]: ...
-
-
-def docker_boundary_command(
-    *,
-    image: str,
-    workspace: Path,
-    cidfile: Path | None = None,
-    container_name: str | None = None,
-    platform: str | None = None,
-) -> list[str]:
-    """Return the single hardened Docker boundary shared by all runners."""
-
-    command = [
-        "docker",
-        "run",
-        "-i",
-        "--rm",
-        "--network",
-        "none",
-        "--user",
-        "runner",
-        "--read-only",
-        "--cap-drop",
-        "ALL",
-        "--security-opt",
-        "no-new-privileges",
-        "--pids-limit",
-        "128",
-        "--memory",
-        "256m",
-        "--cpus",
-        "1",
-        "-v",
-        f"{workspace.resolve()}:/submission:ro",
-        "--tmpfs",
-        "/thebitlab-work:rw,exec,nosuid,nodev,mode=1777,size=64m",
-        "-e",
-        "TMPDIR=/thebitlab-work",
-        "-w",
-        "/submission",
-    ]
-    if platform is not None:
-        command.extend(["--platform", platform])
-    if cidfile is not None:
-        command.extend(["--cidfile", str(cidfile.resolve())])
-    if container_name is not None:
-        command.extend(["--name", container_name])
-    command.append(image)
-    return command
 
 
 class DockerRuntimeSandboxExecutionService:
